created class "LogisticModelROCCurve" for calculating ROC curve by outputting the probabilities of true positive for each individual word used for classification

+++ src/main/java/se/kth/speech/coin/tangrams/data/Referent.java
@@ -28,7 +28,7 @@ public class Referent {
 	public float midx = 0f;
 	public float midy = 0f;
 	
-	public float mentioned = 0f;
+	public int mentioned = 0;
 	
 	public static Set<String> shapes = new HashSet<>((int) Math.ceil(DatasetConstants.EXPECTED_UNIQUE_SHAPE_COUNT / 0.75f));
